+++ bootstrap/app.php
@@ -11,7 +11,10 @@ return Application::configure(basePath: dirname(__DIR__))
         health: '/up',
     )
     ->withMiddleware(function (Middleware $middleware) {
-        //
+        // Route middleware aliases
+        $middleware->alias([
+            'resolve.country' => \App\Http\Middleware\ResolveCountryMiddleware::class,
+        ]);
     })
     ->withExceptions(function (Exceptions $exceptions) {
         //